Split holds tests into their own class.
